refactor: move axi pipeline to codegen
